poverty-app/requirements.txt
Flask
keras==2.2.5
pillow
tensorflow==1.14.0
numpy